Moving routing integration tests to routing directory.
