+++ routing/routing_integration_tests/osrm_route_test.cpp
@@ -1,8 +1,8 @@
 #include "testing/testing.hpp"
 
-#include "integration_tests/routing_test_tools.hpp"
+#include "routing/routing_integration_tests/routing_test_tools.hpp"
 
-#include "../indexer/mercator.hpp"
+#include "../../indexer/mercator.hpp"
 
 using namespace routing;
