+++ routing/routing_integration_tests/pedestrian_route_test.cpp
@@ -1,6 +1,6 @@
 #include "testing/testing.hpp"
 
-#include "integration_tests/routing_test_tools.hpp"
+#include "routing/routing_integration_tests/routing_test_tools.hpp"
 
 #include "indexer/mercator.hpp"
