+++ routing/routing_integration_tests/routing_integration_tests.pro
@@ -5,12 +5,12 @@
 # - tests are written to be launch on the whole world dataset;
 # - tests covers significant number of subsystems;
 
-TARGET = integration_tests
+TARGET = routing_integration_tests
 CONFIG += console warn_on
 CONFIG -= app_bundle
 TEMPLATE = app
 
-ROOT_DIR = ..
+ROOT_DIR = ../..
 DEPENDENCIES = map routing search storage indexer platform geometry coding base osrm jansson protobuf tomcrypt succinct stats_client
 
 macx-*: LIBS *= "-framework IOKit" "-framework SystemConfiguration"
@@ -20,7 +20,7 @@ include($$ROOT_DIR/common.pri)
 QT *= core
 
 SOURCES += \
-  ../testing/testingmain.cpp \
+  ../../testing/testingmain.cpp \
   online_cross_tests.cpp \
   osrm_route_test.cpp \
   osrm_turn_test.cpp \
