+++ routing/routing_integration_tests/routing_test_tools.cpp
@@ -1,4 +1,4 @@
-#include "integration_tests/routing_test_tools.hpp"
+#include "routing/routing_integration_tests/routing_test_tools.hpp"
 
 #include "testing/testing.hpp"
